pyproject.toml
[tool.poetry]
name = "dagster-etl"
version = "0.1.0"
description = ""
authors = ["Marcos Alcozer <marcos@alcozer.dev>"]
license = "Apache License"

[tool.poetry.dependencies]
python = "3.9.10"
requests = "^2.27.1"
black = "^22.1.0"
dagster = "^0.14.1"
dagster-dbt = "^0.14.1"
dagster-gcp = "^0.14.1"
dagit = "^0.14.1"
dbt-bigquery = "^1.0.0"
google-api-python-client = "1.12.10"
google-cloud-storage = "^2.1.0"
pandas = "^1.4.1"
pyarrow = "^7.0.0"
tenacity = "^8.0.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
